+++ src/autocomplete/src/autocomplete.js
@@ -1,0 +1,11 @@
+/**
+ * export autocomplete
+ * @author yiminghe@gmail.com
+ */
+KISSY.add("autocomplete", function (S, AutoComplete, Input, LocalDataSource) {
+    AutoComplete.Input = Input;
+    AutoComplete.LocalDataSource = LocalDataSource;
+    return AutoComplete;
+}, {
+    requires:['autocomplete/base', 'autocomplete/input', 'autocomplete/localdatasource']
+})