+++ src/autocomplete/src/remotedata.js
@@ -1,0 +1,7 @@
+/**
+ * remote datasource
+ * @author yiminghe@gmail.com
+ */
+KISSY.add("autocomplete/remotedata", function () {
+
+});