+++ src/menu/demo/delmenuitemrender.js
@@ -2,9 +2,10 @@
  * @fileOverview deletable menuitemrender
  * @author yiminghe@gmail.com
  */
-KISSY.add("menu/delmenuitemrender", function(S, Node, UIBase, Component, MenuItemRender) {
+KISSY.add("menu/delmenuitemrender", function (S, Node, UIBase, Component, Menu) {
     var CLS = "menuitem-deletable",
         DEL_CLS = "menuitem-delete";
+    var MenuItemRender = Menu.Item.Render;
     var DEL_TMPL = '<span class="{prefixCls}' + DEL_CLS + '" title="{tooltip}">X<' + '/span>';
 
     function addDel(self) {
@@ -15,16 +16,16 @@ KISSY.add("menu/delmenuitemrender", function(S, Node, UIBase, Component, MenuIte
     }
 
     return UIBase.create(MenuItemRender, {
-        createDom:function() {
+        createDom:function () {
             addDel(this);
         },
-        _uiSetContent:function(v) {
+        _uiSetContent:function (v) {
             var self = this;
             MenuItemRender.prototype._uiSetContent.call(self, v);
             addDel(self);
         },
 
-        _uiSetDelTooltip:function() {
+        _uiSetDelTooltip:function () {
             this._uiSetContent(this.get("content"));
         }
     }, {
@@ -32,7 +33,7 @@ KISSY.add("menu/delmenuitemrender", function(S, Node, UIBase, Component, MenuIte
             delTooltip:{}
         },
         HTML_PARSER:{
-            delEl:function(el) {
+            delEl:function (el) {
                 return el.one(this.getCls(DEL_CLS));
             }
         },
@@ -40,5 +41,5 @@ KISSY.add("menu/delmenuitemrender", function(S, Node, UIBase, Component, MenuIte
         DEL_CLS:DEL_CLS
     });
 }, {
-    requires:['node','uibase','component','./menuitemrender']
+    requires:['node', 'uibase', 'component', 'menu']
 });