+++ src/menu/src/menu.js
@@ -2,7 +2,7 @@
  * @fileOverview menu
  * @author yiminghe@gmail.com
  */
-KISSY.add("menu", function(S, Menu, Render, Item, ItemRender, SubMenu, SubMenuRender, Separator, SeparatorRender, PopupMenu, FilterMenu, DelMenuItem) {
+KISSY.add("menu", function (S, Menu, Render, Item, ItemRender, SubMenu, SubMenuRender, Separator, SeparatorRender, PopupMenu, FilterMenu) {
     Menu.Render = Render;
     Menu.Item = Item;
     Menu.Item.Render = ItemRender;
@@ -11,7 +11,6 @@ KISSY.add("menu", function(S, Menu, Render, Item, ItemRender, SubMenu, SubMenuRe
     Menu.Separator = Separator;
     Menu.PopupMenu = PopupMenu;
     Menu.FilterMenu = FilterMenu;
-    Menu.DelMenuItem = DelMenuItem;
     return Menu;
 }, {
     requires:[
@@ -24,8 +23,6 @@ KISSY.add("menu", function(S, Menu, Render, Item, ItemRender, SubMenu, SubMenuRe
         'menu/separator',
         'menu/separatorrender',
         'menu/popupmenu',
-        'menu/filtermenu',
-        'menu/delmenuitem',
-        'menu/delmenuitemrender'
+        'menu/filtermenu'
     ]
 });